refactor(2198): move the Agent version floors out of agent.lib.mjs after the merge
